In the Auth class, add an option to export the login() method so that it can be called from vanilla JS.
Add a sample usage of this pattern, called through JSNI, to the sample
page.

Add a module that exports the login() method on module load, and
compiles down to a single script. This allows for the GWT library to be
packaged and released as a standalone JS library. It currently compiles
down to ~14KB.

R=cromwellian
DELTA=472  (430 added, 12 deleted, 30 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3293